util/hashx: move from sha256x
